AEA-3830 Consolidated check digit calculation.
